requirements.txt
certifi==2023.7.22 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
emoji==2.8.0 ; python_version >= "3.10" and python_version < "4.0"
filelock==3.12.2 ; python_version >= "3.10" and python_version < "4.0"
fsspec==2023.6.0 ; python_version >= "3.10" and python_version < "4.0"
huggingface-hub==0.16.4 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "4.0"
mpmath==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
networkx==3.1 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.25.2 ; python_version < "4.0" and python_version >= "3.10"
packaging==23.1 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.0.3 ; python_version >= "3.10" and python_version < "4.0"
pillow==10.0.0 ; python_version >= "3.10" and python_version < "4.0"
pyairtable==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.12 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
regex==2023.8.8 ; python_version >= "3.10" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
safetensors==0.3.2 ; python_version >= "3.10" and python_version < "4.0"
schedule==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sympy==1.12 ; python_version >= "3.10" and python_version < "4.0"
tokenizers==0.13.3 ; python_version >= "3.10" and python_version < "4.0"
torch==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
torchaudio==2.0.2 ; python_version >= "3.10" and python_version < "4.0"
torchvision==0.15.2 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.66.1 ; python_version >= "3.10" and python_version < "4.0"
transformers==4.31.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "4.0"
urllib3==2.0.4 ; python_version >= "3.10" and python_version < "4.0"
twscrape==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
